% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{align_kmers}
\alias{align_kmers}
\title{Align k-mers in a dataframe}
\usage{
align_kmers(df, is_log2 = FALSE, limit = Inf)
}
\arguments{
\item{df}{Dataframe containing k-mers}

\item{is_log2}{Get 2^enrichments}

\item{limit}{Limit the number of k-mers to align}
}
\value{
Dataframe with aligned k-mers
}
\description{
Align k-mers in a dataframe
}
\examples{
# Load sample data
data(rbfox2_enrichments)

# Align all k-mers in data.frame
aligned <- align_kmers(rbfox2_enrichments)
head(aligned)

# Align only the first five k-mers in data.frame
aligned <- align_kmers(rbfox2_enrichments, limit = 5)
head(aligned)
}
