% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmers.R
\name{enrichments}
\alias{enrichments}
\title{Calculate k-mer enrichments}
\usage{
enrichments(
  testfile,
  ctrlfile = NULL,
  kmer = 3,
  probabilistic = FALSE,
  normalize = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{testfile}{Test sequences}

\item{ctrlfile}{Control sequences}

\item{kmer}{Length of kmer}

\item{probabilistic}{Compute probabilistic enrichments}

\item{normalize}{Get log2 of enrichments}

\item{verbose}{Verbose output for calculations}
}
\value{
dataframe containing enrichments
}
\description{
Calculate k-mer enrichments
}
\examples{
print("Hello, world!")
}
