% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_pwms}
\alias{get_pwms}
\title{Get the positional weight matrix (PWM) from enriched k-mers}
\usage{
get_pwms(
  df,
  num_pwms = 1,
  limit = Inf,
  limit_per_logo = Inf,
  type = c("DNA", "RNA")
)
}
\arguments{
\item{df}{Dataframe containing k-mers and associated enrichments. Meant to
complement enrichments/ikke function, so use their output.}

\item{num_pwms}{Number of PWMs to generate from \code{df}}

\item{limit}{Number of rows to use from \code{df}}

\item{limit_per_logo}{Number of rows to use per logo}

\item{type}{Sequence type. Can be one of either 'DNA' or 'RNA'}
}
\value{
List of PWMs, containing 'n', the number of kmers in the pwm, and
'pwm', the values of the PWM.
}
\description{
get_pwms is a simple method to compute the PWMs from k-mer enrichments. To
accomplish this, k-mers are first aligned to the most enriched k-mer. K-mers
that do not align to the most enriched k-mer are then used to create another
PWM. Weight is determined as the enrichment of the k-mer, non significant
positions are trimmed, and the PWM is normalized per position.
}
\examples{
# Load sample data
data(rbfox2_enrichments)

# Get the PWM for the enrichments
pwm <- get_pwms(rbfox2_enrichments)

# Get the primary and secondary PWMs
pwm <- get_pwms(rbfox2_enrichments, num_pwms = 2)

# Limit the total number of enrichments to use in PWM
pwm <- get_pwms(rbfox2_enrichments, limit = 20)

# Use 'U' instead of 'T' for RNA-based PWM
pwm <- get_pwms(rbfox2_enrichments, type = "RNA")
}
