% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{hamming_distance}
\alias{hamming_distance}
\title{Hamming distance of two strings}
\usage{
hamming_distance(str1, str2)
}
\arguments{
\item{str1}{First string}

\item{str2}{Second string}
}
\value{
Hamming distance between both strings
}
\description{
Hamming distance of two strings
}
