% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/katss.R
\name{ikke}
\alias{ikke}
\title{Title Iterative K-mer Knockout Enrichments}
\usage{
ikke(
  testfile,
  ctrlfile = NULL,
  kmer = 3,
  iterations = 10,
  probabilistic = FALSE,
  normalize = FALSE,
  threads = 1
)
}
\arguments{
\item{testfile}{Test sequences file. Can be in FASTQ, FASTA, or raw sequences
format. Raw sequences format is a file containing only "A", "C", "G", and "T"
/"U" characters, in every sequence separated by newline.}

\item{ctrlfile}{Control sequences file. Can be in FASTQ, FASTA, or raw sequences
format. Raw sequences format is a file containing only "A", "C", "G", and "T"
/"U" characters, in every sequence separated by newline.}

\item{kmer}{Length of k-mer.}

\item{iterations}{Number of iterations to perform}

\item{probabilistic}{Calculate probabilistic enrichments.}

\item{normalize}{Normalize enrichments to log2}

\item{threads}{Number of threads to use. Specifying less than 1 thread
sets the number of threads as 1.}
}
\value{
data.frame containing the enrichments
}
\description{
Title Iterative K-mer Knockout Enrichments
}
\examples{
# Load data
data(rbfox2_seqs)

test_seqs <- tempfile()
writeLines(rbfox2_seqs$bound, test_seqs)

# Get the enrichments without a control
result <- ikke(test_seqs, probabilistic = TRUE)
head(result)

# Get the 5-mer enrichments
result <- ikke(test_seqs, kmer = 5, probabilistic = TRUE)
head(result)

ctrl_seqs <- tempfile()
writeLines(rbfox2_seqs$input, ctrl_seqs)

# Get the enrichments when you have a control
result <- ikke(test_seqs, ctrl_seqs, kmer = 5)
head(result)

# Specify the number of enrichments to obtain
result <- ikke(test_seqs, ctrl_seqs, kmer = 5, iterations = 2)
print(result)

# Normalize enrichments to log2
result <- ikke(test_seqs, ctrl_seqs, kmer = 5, normalize = TRUE)
head(result)
tail(result)
}
