% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pad_kmer}
\alias{pad_kmer}
\title{pad a kmer string}
\usage{
pad_kmer(kmer, max_lpad, max_rpad)
}
\arguments{
\item{kmer}{String to pad}

\item{max_lpad}{Amount to pad to the left}

\item{max_rpad}{Amount to pad to the right}
}
\value{
Padded string \code{kmer}
}
\description{
pad a kmer string
}
