% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_logo}
\alias{plot_logo}
\title{Plot sequence logo from a PWM}
\usage{
plot_logo(
  logo_list,
  ncol = NULL,
  method = c("bits", "probability"),
  title = "Weighted Sequence Logo",
  name = NULL,
  subtitle = NULL
)
}
\arguments{
\item{logo_list}{List of logos you want to plot}

\item{ncol}{Number of columns}

\item{method}{Use bits of probability method for sequence logo.}

\item{title}{Title of the plot}

\item{name}{Set name of y axis.}

\item{subtitle}{Title for each PWM. Should be same length as the number of PWMs.}
}
\value{
ggplot object of PWM plot
}
\description{
Plots PWMs generated from \code{\link{get_pwms}}. This is a simple interface
over the ggseqlogo package. For more complex plots or greater customization,
see \link[ggseqlogo]{ggseqlogo} and \link[ggseqlogo]{geom_logo}
}
\examples{
# Load sample data
data(rbfox2_pwms)

# Plot multiple logos
rbfox2_plot <- plot_logo(rbfox2_pwms)
print(rbfox2_plot)

# Plot multiple logos using probability
rbfox2_plot <- plot_logo(rbfox2_pwms, method = "probability")
print(rbfox2_plot)
}
