% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{rbfox2_enrichments}
\alias{rbfox2_enrichments}
\title{Data frame containing the 5-mer enrichments of RNA-binding protein RBFOX2}
\format{
An object of class \code{data.frame} with 1024 rows and 2 columns.
}
\usage{
data(rbfox2_enrichments)
}
\description{
Data frame containing the 5-mer enrichments of RNA-binding protein RBFOX2
}
\keyword{data}
