% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{rbfox2_pwms}
\alias{rbfox2_pwms}
\title{List of matrices containing the 5-mer positional weight matrix of RNA-binding
protein RBFOX2}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(rbfox2_pwms)
}
\description{
List of matrices containing the 5-mer positional weight matrix of RNA-binding
protein RBFOX2
}
\keyword{data}
