% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/katss.R
\docType{data}
\name{rbfox2_seqs}
\alias{rbfox2_seqs}
\title{List containing the test and control sequences for RNA-binding protein RBFOX2}
\format{
A list containing two lists: a list of control sequences, and a list
of test sequences.
\describe{
\item{input}{1000 control sequences of 20 nt long}
\item{bound}{1000 test sequences of 20 nt long}
}
}
\usage{
data(rbfox2_seqs)
}
\description{
List containing the test and control sequences for RNA-binding protein RBFOX2
}
\keyword{data}
