% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkats-package.R
\docType{package}
\name{rkats-package}
\alias{rkats}
\alias{rkats-package}
\title{rkats: R wrapper for the KATSS package}
\description{
Analyze RNA sequences and identify enriched motifs in protein-bound RNA. The package comes with functions for k-mer counting, enrichment calculations, k-mer alignment, and plotting utilities.
}
\author{
\strong{Maintainer}: Francisco F Cavazos \email{francisco.cavazos03@gmail.com} (\href{https://orcid.org/0009-0000-1804-9601}{ORCID})

Authors:
\itemize{
  \item Yue Hu
  \item Daniel Dominguez
}

}
\keyword{internal}
