% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkatss-package.R
\docType{package}
\name{rkatss-package}
\alias{rkatss}
\alias{rkatss-package}
\title{rkatss: R wrapper for the KATSS package}
\description{
A suite of tools for k-mer based analysis of sequence. Tools such as, streaming based k-mer counting and enrichment analysis, motif identification and plotting, and more.
}
\author{
\strong{Maintainer}: Francisco Cavazos \email{francisco.cavazos03@gmail.com} (\href{https://orcid.org/0009-0000-1804-9601}{ORCID})

}
\keyword{internal}
