% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmers.R
\name{seqseq}
\alias{seqseq}
\title{Search for a sequence within a sequence}
\usage{
seqseq(sequence, search, all.matches = TRUE)
}
\arguments{
\item{sequence}{Big sequence}

\item{search}{Small sequence to search for}
}
\value{
Index in which search was found, 0 if not found
}
\description{
This function returns the index in which a sub-sequence is found within a
larger sequence, or 0 if not found. Searching is case insensitive and 'T' and
'U' are equivalent. The maximum length sequence to search for
is currently capped at 255.
}
\examples{
## Find "AAA" in a random sequence
seq <-paste(sample(c("A","C","G","T"), 1000, TRUE), collapse="")
seqseq(seq, "AAA")

## Searching is case-insensitive
seq <- "accgtaagggtgccttac"
find2 <- "GGGT"
seqseq(seq, "GGGT")
seqseq(seq, "gCCtT")

## T and U are both interchangeable
seqseq(seq, "GCCUU")

## 0 is returned when search is not in sequence
seqseq(seq, "AAA")
}
