% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/katss.R
\name{count_kmers}
\alias{count_kmers}
\title{K-mer counting}
\usage{
count_kmers(file, kmer = 3)
}
\arguments{
\item{file}{Name of the file which you want to count k-mers from
The file has to be of either: raw sequences, fasta, or fastq format. Other
file types are currently unsupported and will not work properly if used.}

\item{kmer}{Length of the k-mer you want to count. Currently, only k-mers up
to length 16 are supported.}
}
\value{
Dataframe containing the counts for all k-mers
}
\description{
Count the k-mers in a fastq, fasta, or raw sequences file.
}
\examples{
# Create temporary file with sequences
set.seed(1)
test1 <- do.call(paste0, replicate(50, sample(c("A","C","G","T"), 1000, TRUE), FALSE))
tf <- tempfile()
writeLines(test1, tf)

# Count di-nucleotides in file
count_kmers(tf, kmer = 2)

# Count mono-nucleotide in file
count_kmers(tf, kmer = 1)
unlink(tf)
}
