% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/katss.R
\name{enrichments}
\alias{enrichments}
\title{Calculate k-mer enrichments}
\usage{
enrichments(
  testfile,
  ctrlfile = NULL,
  kmer = 3,
  probabilistic = FALSE,
  normalize = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{testfile}{Test sequences}

\item{ctrlfile}{Control sequences}

\item{kmer}{Length of kmer}

\item{probabilistic}{Compute probabilistic enrichments}

\item{normalize}{Get log2 of enrichments}

\item{verbose}{Verbose output for calculations}
}
\value{
dataframe containing enrichments
}
\description{
Calculate k-mer enrichments
}
\examples{
# Load data
data(rbfox2_seqs)

test_seqs <- tempfile()
writeLines(rbfox2_seqs$bound, test_seqs)

# Get the enrichments without a control
result <- enrichments(test_seqs, probabilistic = TRUE)
head(result)

# Get the 5-mer enrichments
result <- enrichments(test_seqs, kmer = 5, probabilistic = TRUE)
head(result)

ctrl_seqs <- tempfile()
writeLines(rbfox2_seqs$input, ctrl_seqs)

# Get the enrichments when you have a control
result <- enrichments(test_seqs, ctrl_seqs, kmer = 5)
head(result)

# Normalize enrichments to log2
result <- enrichments(test_seqs, ctrl_seqs, kmer = 5, normalize = TRUE)
head(result)
tail(result)

unlink(test_seqs)
unlink(ctrl_seqs)
}
