% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmers.R
\name{enrichments}
\alias{enrichments}
\title{Calculate enrichments}
\usage{
enrichments(testfile, ctrlfile, kmer, normalize = FALSE)
}
\arguments{
\item{testfile}{Test sequences}

\item{ctrlfile}{Control sequences}

\item{kmer}{Length of kmer}

\item{normalize}{Get log2 of enrichments}
}
\value{
dataframe containing enrichments
}
\description{
Calculate enrichments
}
\examples{
print("Hello, world!")
}
