% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmers.R
\name{ikke}
\alias{ikke}
\title{Title Iterative K-mer Knockout Enrichments}
\usage{
ikke(testfile, ctrlfile, kmer, iterations = 10, normalize = FALSE, threads = 1)
}
\arguments{
\item{testfile}{Test sequences file. Can be in FASTQ, FASTA, or Sequences format}

\item{ctrlfile}{Control sequences file}

\item{kmer}{Length of k-mer}

\item{iterations}{Number of iterations to perform}

\item{normalize}{Normalize enrichments to log2}

\item{threads}{Number of threads to use}
}
\value{
data.frame containing the enrichments
}
\description{
Title Iterative K-mer Knockout Enrichments
}
