% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkatss-package.R
\docType{package}
\name{rkatss-package}
\alias{rkatss}
\alias{rkatss-package}
\title{rkatss: R wrapper for the KATSS package}
\description{
Analyze RNA sequences and identify enriched motifs in protein-bound RNA. The package comes with functions for k-mer counting, enrichment calculations, k-mer alignment, and plotting utilities.
}
\author{
\strong{Maintainer}: Francisco Cavazos \email{francisco.cavazos03@gmail.com} (\href{https://orcid.org/0009-0000-1804-9601}{ORCID})

}
\keyword{internal}
